+++ my-app/src/App.jsx
@@ -1,6 +1,7 @@
 import "./App.css";
 import { BrowserRouter, Route, Routes } from "react-router-dom";
 import Home from "./pages/Home";
+import Login from "./pages/Login";
 import SignUpFirst from "./pages/SignUpFirst";
 import SignUpSecond from "./pages/SignUpSecond";
 
@@ -9,6 +10,7 @@ function App() {
     <BrowserRouter>
       <Routes>
         <Route path="/" element={<Home />} />
+        <Route path="/login" element={<Login />} />
         <Route path="/signup/1" element={<SignUpFirst />} />
         <Route path="/signup/2" element={<SignUpSecond />} />
       </Routes>
